requirements.txt
hydra-core
jinja2
lmstudio
omegaconf
openai
pandas
pydantic
PyYAML
tqdm
